Updated docs/test for dot and vdot
[ghstack-poisoned]
