+++ torch/_tensor_docs.py
@@ -1210,7 +1210,7 @@ In-place version of :meth:`~Tensor.divide`
 
 add_docstr_all('dot',
                r"""
-dot(tensor2) -> Tensor
+dot(other) -> Tensor
 
 See :func:`torch.dot`
 """)
@@ -4004,7 +4004,7 @@ See :func:`torch.var`
 
 add_docstr_all('vdot',
                r"""
-dot(other) -> Tensor
+vdot(other) -> Tensor
 
 See :func:`torch.vdot`
 """)
